Refactor config in launcher
