+++ engine/launcher/src/main/scala/org/enso/launcher/config/GlobalConfiguration.scala
@@ -1,0 +1,3 @@
+package org.enso.launcher.config
+
+class GlobalConfiguration {}
