+++ engine/launcher/src/main/scala/org/enso/launcher/installation/DistributionInstaller.scala
@@ -5,7 +5,7 @@ import java.nio.file.Files
 import org.enso.cli.CLIOutput
 import org.enso.launcher.FileSystem.PathSyntax
 import org.enso.launcher.cli.InternalOpts
-import org.enso.launcher.{FileSystem, GlobalConfigurationManager, Logger, OS}
+import org.enso.launcher.{FileSystem, Logger, OS}
 
 import scala.util.control.NonFatal
