+++ engine/launcher/src/main/scala/org/enso/launcher/project/Project.scala
@@ -4,7 +4,7 @@ import java.io.File
 import java.nio.file.Path
 
 import nl.gn0s1s.bump.SemVer
-import org.enso.launcher.GlobalConfigurationManager
+import org.enso.launcher.config.GlobalConfigurationManager
 import org.enso.pkg.{DefaultEnsoVersion, Package, SemVerEnsoVersion}
 
 /**
