+++ engine/launcher/src/test/scala/org/enso/launcher/components/runner/RunnerSpec.scala
@@ -5,8 +5,9 @@ import java.util.UUID
 
 import nl.gn0s1s.bump.SemVer
 import org.enso.launcher.FileSystem.PathSyntax
-import org.enso.launcher.{GlobalConfigurationManager, Logger}
+import org.enso.launcher.Logger
 import org.enso.launcher.components.ComponentsManagerTest
+import org.enso.launcher.config.GlobalConfigurationManager
 import org.enso.launcher.project.ProjectManager
 
 class RunnerSpec extends ComponentsManagerTest {
